+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 -------------------------
 * Update Manney et al. 2011 method to include 'check_diagonals' parameter
 * Update Manney et al. 2011 method to select maximum windspeed if no cores found above threshold
+* Add 'slice_array_by_index_breaks' to data utils
 * Update examples of use and quick start table in docs